Refactors utils module.
